requirements.txt
torch
pillow
numpy
torchvision
tqdm
h5py
python-dotenv
matplotlib